fix(Order): Remove id from table order_products

+++ src/main/java/com/dulcenectar/java/services/interfaces/OrderService.java
@@ -7,7 +7,7 @@ import com.dulcenectar.java.dtos.order.GetOrderResponseDto;
 
 
 public interface OrderService {
-	public Long createNewOrder(CreateOrderRequestDto order);
+	public Integer createNewOrder(CreateOrderRequestDto order);
 	public List<GetOrderResponseDto> getOrders();
 	public String deleteOrder(Integer orderId);
 }
